fix(api): remove several useless api
