+++ .dockerignore
@@ -1,0 +1,16 @@
+node_modules
+npm-debug.log
+Dockerfile*
+docker-compose*
+.dockerignore
+.git
+.gitignore
+.env
+*/bin
+*/obj
+README.md
+.vscode
+.idea/
+go.sum
+res/
+!res/docker*
