+++ bin/test-attribution-txt.sh
@@ -45,7 +45,7 @@ if [ ! -f Attribution.txt ]; then
 else
     # loop over every library in the modules.txt file in vendor
     while IFS= read -r lib; do
-        if ! grep -q "$lib" Attribution.txt; then
+        if ! grep -q "$lib" Attribution.txt && [ "$lib" != "explicit" ]; then
             echo "An attribution for $lib is missing from in Attribution.txt, please add"
             # need to do this in a bash subshell, see SC2031
             (( EXIT_CODE=1 ))
