+++ cmd/main.go
@@ -1,16 +1,15 @@
 // -*- Mode: Go; indent-tabs-mode: t -*-
 //
-// Copyright (C) 2017-2018 Canonical Ltd
-// Copyright (C) 2018-2019 IOTech Ltd
+// Copyright (C) 2019-2020 Jiangxing Intelligence Ltd
 //
 // SPDX-License-Identifier: Apache-2.0
 
-// This package provides device service  of a GPIO device service.
+// This package provides device service of a GPIO devices.
 package main
 
 import (
 	"github.com/edgexfoundry/device-gpio-go"
-	"github.com/edgexfoundry/device-gpio-go/driver"
+	"github.com/edgexfoundry/device-gpio-go/internal/driver"
 	"github.com/edgexfoundry/device-sdk-go/pkg/startup"
 )
 
@@ -19,7 +18,6 @@ const (
 )
 
 func main() {
-
 	d := driver.Driver{}
 	startup.Bootstrap(serviceName, device.Version, &d)
 }
