+++ cmd/res/device-gpio65-profile.yaml
@@ -1,0 +1,52 @@
+name: "device-gpio65"
+manufacturer: "Jiangxing Intelligence"
+model: "SP-01"
+labels:
+  - "gpio65"
+description: "automatically export gpio65 by command"
+
+deviceResources:
+  -
+    name: "value"
+    description: "set or get system gpio value"
+    properties:
+      value:
+        { type: "Int8", readWrite: "RW", minimum: "0", maximum: "1", defaultValue: "0" }
+      units:
+        { type: "String", readWrite: "R", defaultValue: "high:1; low:0" }
+
+
+deviceCommands:
+  -
+    name: "value"
+    get:
+      - { operation: "get", deviceResource: "value" }
+    set:
+      - { operation: "set", deviceResource: "value", parameter: "0" }
+
+
+coreCommands:
+  -
+    name: "value"
+    put:
+      path: "/api/v1/device/{deviceId}/value"
+      parameterNames: ["value"]
+      responses:
+        -
+          code: "200"
+          description: ""
+        -
+          code: "500"
+          description: "service unavailable"
+          expectedValues: []
+    get:
+      path: "/api/v1/device/{deviceId}/value"
+      responses:
+        -
+          code: "200"
+          description: ""
+          expectedValues: ["value"]
+        -
+          code: "500"
+          description: "service unavailable"
+          expectedValues: []
