+++ docker-compose.yml
@@ -1,15 +1,18 @@
-version: '3'
+version: '3.4'
 
 services:
-  edgex-device-gpio:
+  device-gpio:
     image: edgexfoundry/docker-device-gpio-go:1.0.0-dev
     ports:
-      - "49960:49960"
-    container_name: edgex-device-gpio-go
-    hostname: edgex-device-gpio-go
+      - "0.0.0.0:49950:49950"
+    container_name: edgex-device-gpio
+    hostname: edgex-device-gpio
     networks:
       - edgex-network
-    restart: always
+    environment:
+      <<: *common-variables
+      Service_Host: edgex-device-gpio
+    # restart: always
     privileged: true
     volumes:
       - "/sys/class/gpio:/sys/class/gpio"
