+++ go.mod
@@ -1,13 +1,12 @@
 module github.com/edgexfoundry/device-gpio-go
 
 require (
-	github.com/edgexfoundry/device-sdk-go v1.2.3
-	github.com/edgexfoundry/go-mod-core-contracts v0.1.58
-	github.com/google/uuid v1.1.1 // indirect
+	github.com/edgexfoundry/device-sdk-go v1.4.0
+	github.com/edgexfoundry/go-mod-core-contracts v0.1.115
 	github.com/hashicorp/go-sockaddr v1.0.1 // indirect
 	github.com/hashicorp/golang-lru v0.5.1 // indirect
 	github.com/miekg/dns v1.1.4 // indirect
 	github.com/mitchellh/go-homedir v1.1.0 // indirect
 )
 
-go 1.13
+go 1.15
